mss_ai_ppt_sample_assets/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-pptx>=0.6.23
jsonschema>=4.20.0
